# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  authzapi.createMenuProfileBody:
    properties:
      description:
        type: string
      menus:
        description: '["kcontrol","dashboard",...]'
        items:
          type: string
        type: array
      name:
        type: string
      relations:
        description: '["viewer","creator",...] default ["viewer"]'
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
...
  authzapi.updateMenuProfileBody:
    properties:
      description:
        type: string
      menus:
        items:
          type: string
        type: array
      name:
        type: string
      orgUnits:
        description: nil = not provided (keep old), [] = clear/reset, [...] = set
        items:
          type: string
        type: array
      relations:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
...
  gmod.ApiErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details: {}
      message:
        example: Invalid request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  gmod.PaginatedResponse:
    properties:
      code:
        example: SUCCESS
        type: string
      details: {}
      message:
        example: profiles fetched
        type: string
      pagination:
        $ref: '#/definitions/gmod.Pagination'
      status:
        example: true
        type: boolean
    type: object
  gmod.Pagination:
    properties:
      page:
        type: integer
      perPages:
        type: integer
      sortField:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
...
  gmod.SuccessDetailResponseAny:
    properties:
      details: {}
      status:
        type: boolean
    type: object
...
  gmod.SuccessDetailResponseMenuAccess:
    properties:
      code:
        example: SUCCESS
        type: string
      details: {}
      message:
        example: accessible menus fetched
        type: string
      status:
        example: true
        type: boolean
    type: object
  gmod.SuccessDetailResponseResourceAccess:
    properties:
      code:
        example: SUCCESS
        type: string
      details: {}
      message:
        example: accessible cameras fetched
        type: string
      status:
        example: true
        type: boolean
    type: object
...
  gmod.SuccessMessageResponse:
    properties:
      code:
        example: SUCCESS
        type: string
      message:
        example: Operation successful
        type: string
      status:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: REST API for Gateway system
  title: Gateway API
  version: "1.0"
paths:
...
  /orgs/menu/access:
    get:
      description: Returns menu IDs the caller can access via their OrgUnit membership
        and active MenuPermissionProfiles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseMenuAccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List menus accessible to the current user
      tags:
      - MemberAccess
  /orgs/menu/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List all available menus
      tags:
      - MenuPermissionProfile
  /orgs/menu/permissions:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: items per page (default 10)
        in: query
        name: perPages
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      - description: sort field (default createdAt)
        in: query
        name: sortField
        type: string
      - description: 'sort order: asc|desc (default desc)'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List menu permission profiles
      tags:
      - MenuPermissionProfile
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authzapi.createMenuProfileBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Create menu permission profile
      tags:
      - MenuPermissionProfile
  /orgs/menu/permissions/{id}:
    delete:
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete menu permission profile
      tags:
      - MenuPermissionProfile
    get:
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get menu permission profile by ID
      tags:
      - MenuPermissionProfile
    patch:
      consumes:
      - application/json
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authzapi.updateMenuProfileBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update menu permission profile
      tags:
      - MenuPermissionProfile
  /orgs/resource/access:
    get:
      description: Returns paginated cameras the caller can access via OrgUnit membership
        and active ResourcePermissionProfiles, enriched with the union of Permify
        relations (viewer, editor, …).
      parameters:
      - description: Filter by resource type (e.g. camera)
        in: query
        name: resourceType
        type: string
      - description: Search camera name
        in: query
        name: search
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: perPages
        type: integer
      - description: Sort field (default createAt)
        in: query
        name: sortField
        type: string
      - description: 'Sort order: asc | desc (default desc)'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseResourceAccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List cameras accessible to the current user
      tags:
      - MemberAccess
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"